+++ src/integrationTest/java/net/sf/jabref/gui/GUITest.java
@@ -1,6 +1,7 @@
 package net.sf.jabref.gui;
 
 import net.sf.jabref.JabRefMain;
+import net.sf.jabref.gui.dbproperties.DatabasePropertiesDialog;
 import net.sf.jabref.gui.preftabs.PreferencesDialog;
 import org.assertj.swing.core.GenericTypeMatcher;
 import org.assertj.swing.dependency.jsr305.Nonnull;
@@ -19,6 +20,7 @@ import static org.assertj.swing.launcher.ApplicationLauncher.application;
 public class GUITest extends AssertJSwingJUnitTestCase {
 
     private AWTExceptionHandler awtExceptionHandler;
+
 
     @Override
     protected void onSetUp() {
@@ -64,12 +66,14 @@ public class GUITest extends AssertJSwingJUnitTestCase {
         newDatabase(mainFrame);
 
         mainFrame.menuItemWithPath("BibTeX", "New entry...").click();
-        findDialog(EntryTypeDialog.class).withTimeout(10_000).using(robot()).button(new GenericTypeMatcher<JButton>(JButton.class) {
+        findDialog(EntryTypeDialog.class).withTimeout(10_000).using(robot())
+                .button(new GenericTypeMatcher<JButton>(JButton.class) {
 
-            @Override protected boolean isMatching(@Nonnull JButton jButton) {
-                return "Book".equals(jButton.getText());
-            }
-        }).click();
+                    @Override
+                    protected boolean isMatching(@Nonnull JButton jButton) {
+                        return "Book".equals(jButton.getText());
+                    }
+                }).click();
 
         exitJabRef(mainFrame);
     }
@@ -82,12 +86,14 @@ public class GUITest extends AssertJSwingJUnitTestCase {
 
         robot().waitForIdle();
 
-        findDialog(PreferencesDialog.class).withTimeout(10_000).using(robot()).button(new GenericTypeMatcher<JButton>(JButton.class) {
+        findDialog(PreferencesDialog.class).withTimeout(10_000).using(robot())
+                .button(new GenericTypeMatcher<JButton>(JButton.class) {
 
-            @Override protected boolean isMatching(@Nonnull JButton jButton) {
-                return "OK".equals(jButton.getText());
-            }
-        }).click();
+                    @Override
+                    protected boolean isMatching(@Nonnull JButton jButton) {
+                        return "OK".equals(jButton.getText());
+                    }
+                }).click();
 
         exitJabRef(mainFrame);
     }
@@ -112,4 +118,25 @@ public class GUITest extends AssertJSwingJUnitTestCase {
         exitJabRef(mainFrame);
     }
 
+    @Test
+    public void testDatabasePropertiesDialog() {
+
+        FrameFixture mainFrame = findFrame(JabRefFrame.class).withTimeout(10_000).using(robot());
+        newDatabase(mainFrame);
+
+        mainFrame.menuItemWithPath("File", "Database properties").click();
+
+        robot().waitForIdle();
+
+        findDialog(DatabasePropertiesDialog.class).withTimeout(10_000).using(robot())
+                .button(new GenericTypeMatcher<JButton>(JButton.class) {
+
+                    @Override
+                    protected boolean isMatching(@Nonnull JButton jButton) {
+                        return "OK".equals(jButton.getText());
+                    }
+                }).click();
+
+        exitJabRef(mainFrame);
+    }
 }